requirements.txt
numpy>=1.21.0
pyyaml>=5.4.0
matplotlib>=3.5.0
mesa>=2.0.0
google-generativeai>=0.3.0

